+++ crates/prql_ast/src/expr.rs
@@ -1,5 +1,7 @@
 mod ident;
 mod literal;
+mod ops;
 
 pub use ident::Ident;
 pub use literal::{Literal, ValueAndUnit};
+pub use ops::{BinOp, UnOp};
