+++ crates/prql_ast/src/expr/ops.rs
@@ -1,0 +1,71 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(
+    Debug,
+    PartialEq,
+    Eq,
+    Clone,
+    Copy,
+    Hash,
+    Serialize,
+    Deserialize,
+    strum::Display,
+    strum::EnumString,
+)]
+pub enum UnOp {
+    #[strum(to_string = "-")]
+    Neg,
+    #[strum(to_string = "+")]
+    Add, // TODO: rename to Pos
+    #[strum(to_string = "!")]
+    Not,
+    #[strum(to_string = "==")]
+    EqSelf,
+}
+
+#[derive(
+    Debug,
+    PartialEq,
+    Eq,
+    Clone,
+    Copy,
+    Hash,
+    Serialize,
+    Deserialize,
+    strum::Display,
+    strum::EnumString,
+)]
+pub enum BinOp {
+    #[strum(to_string = "*")]
+    Mul,
+    #[strum(to_string = "//")]
+    DivInt,
+    #[strum(to_string = "/")]
+    DivFloat,
+    #[strum(to_string = "%")]
+    Mod,
+    #[strum(to_string = "+")]
+    Add,
+    #[strum(to_string = "-")]
+    Sub,
+    #[strum(to_string = "==")]
+    Eq,
+    #[strum(to_string = "!=")]
+    Ne,
+    #[strum(to_string = ">")]
+    Gt,
+    #[strum(to_string = "<")]
+    Lt,
+    #[strum(to_string = ">=")]
+    Gte,
+    #[strum(to_string = "<=")]
+    Lte,
+    #[strum(to_string = "~=")]
+    RegexSearch,
+    #[strum(to_string = "&&")]
+    And,
+    #[strum(to_string = "||")]
+    Or,
+    #[strum(to_string = "??")]
+    Coalesce,
+}
