+++ prql-compiler/src/ast/pl/expr.rs
@@ -10,6 +10,8 @@ use ast::generic as generic_expr;
 
 use super::*;
 
+pub use prql_ast::expr::{BinOp, UnOp};
+
 /// Expr is anything that has a value and thus a type.
 /// If it cannot contain nested Exprs, is should be under [ExprKind::Literal].
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
@@ -98,80 +100,10 @@ pub struct BinaryExpr {
     pub right: Box<Expr>,
 }
 
-#[derive(
-    Debug,
-    PartialEq,
-    Eq,
-    Clone,
-    Copy,
-    Hash,
-    Serialize,
-    Deserialize,
-    strum::Display,
-    strum::EnumString,
-)]
-pub enum BinOp {
-    #[strum(to_string = "*")]
-    Mul,
-    #[strum(to_string = "//")]
-    DivInt,
-    #[strum(to_string = "/")]
-    DivFloat,
-    #[strum(to_string = "%")]
-    Mod,
-    #[strum(to_string = "+")]
-    Add,
-    #[strum(to_string = "-")]
-    Sub,
-    #[strum(to_string = "==")]
-    Eq,
-    #[strum(to_string = "!=")]
-    Ne,
-    #[strum(to_string = ">")]
-    Gt,
-    #[strum(to_string = "<")]
-    Lt,
-    #[strum(to_string = ">=")]
-    Gte,
-    #[strum(to_string = "<=")]
-    Lte,
-    #[strum(to_string = "~=")]
-    RegexSearch,
-    #[strum(to_string = "&&")]
-    And,
-    #[strum(to_string = "||")]
-    Or,
-    #[strum(to_string = "??")]
-    Coalesce,
-}
-
 #[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
 pub struct UnaryExpr {
     pub op: UnOp,
     pub expr: Box<Expr>,
-}
-
-#[derive(
-    Debug,
-    PartialEq,
-    Eq,
-    Clone,
-    Copy,
-    Hash,
-    Serialize,
-    Deserialize,
-    strum::Display,
-    strum::EnumString,
-)]
-pub enum UnOp {
-    #[strum(to_string = "-")]
-    Neg,
-    #[strum(to_string = "+")]
-    Add, // TODO: rename to Pos
-    #[strum(to_string = "!")]
-    Not,
-    #[strum(to_string = "==")]
-    EqSelf,
 }
 
 /// Function call.
